// lib/enums.ts
export enum QUESTION_TYPE {
  MULTIPLE_CHOICE = "MULTIPLE_CHOICE",
  DRAG_AND_DROP = "DRAG_AND_DROP",
  SCENARIO_BASED = "SCENARIO_BASED",
  SHORT_ANSWER = "SHORT_ANSWER",
  FILL_IN_THE_BLANK = "FILL_IN_THE_BLANK",
  MATCHING = "MATCHING",
  TRUE_FALSE = "TRUE_FALSE",
  ORDERING = "ORDERING",
  IMAGE_IDENTIFICATION = "IMAGE_IDENTIFICATION",
}